minor changes
- added an option to change the leaderboard update interval
- moved over the leaderboard format to the config
- fixed a bug

+++ src/main/java/io/tofpu/speedbridge2/domain/player/object/BridgePlayer.java
@@ -78,7 +78,7 @@ public final class BridgePlayer extends CommonBridgePlayer<Player> {
         return score;
     }
 
-    public Score setNewScore(final int islandSlot, final long newScore) {
+    public Score setNewScore(final int islandSlot, final double newScore) {
         final Score score = Score.of(islandSlot, newScore);
         return setNewScore(score);
     }
